Extracted formatDate() into its own utils module
I'm planning on reusing this between a couple components, so it shouldn't
be tied to just FileCell anymore.
